A patch from Gil Portenseigne for OFBIZ-6545 - Replace webtools/control/EntitySQLProcessor FTL with widgets
And while at it set ofbiz.org as default selGroup when none exist. This will prevent the unobvious error message
  <<Unable to esablish a connection with the database. (Unable to acquire a new connection from the pool)>>
in most cases!

This patch replaces EntitySQLCommand ftl to widget. List screen couldn't be migrated to widget due to dynamic column rendering.
To get the default group, a little trick is used, constructing the list of map for drop-down list-option widget.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1692725 13f79535-47bb-0310-9956-ffa450edef68