Fixes an issue crossed while checking OFBIZ-6504 - how to Create PDF with password Protected in Ofbiz
When a description used in a drop-down comes from a localised label in CommonEntityLabels.xml and contains a single quote (') this single quote must be escaped to allow the rendering. Else it's a single quote inside 2 single quotes for the option like:
{'key':'EXST_CANCELLED','description':'Annuler l'exemple (En projet)'}
with he percent encoding there is no issue:
{'key':'EXST_CANCELLED','description':'Annuler l%27exemple (En projet)'}

We noted, with Jinghai, that using html entities does not work. I guess due to another encoding further in the chain. I also tried to percent encoding but was unable to find how to replace a % in Freemarker. Thanks to Gareth's suggestion I finally escaped the quote. Also to escape the quote with replaceAll() in Java you have to put 5 \,  small but weird issues ;) I hope nothing else somewhere already escape the quote, we will see anyway...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1693242 13f79535-47bb-0310-9956-ffa450edef68