At http://www.apache.org/dev/licensing-howto.html#mod-notice, it's adviced:
<<Copyright notifications which have been relocated from source files (rather than removed) must be preserved in NOTICE. However, elements such as the copyright notifications embedded within BSD and MIT licenses need not be duplicated in NOTICE -- it suffices to leave those notices in their original locations.

It is important to keep NOTICE as brief and simple as possible, as each addition places a burden on downstream consumers.

Do not add anything to NOTICE which is not legally required.>>

So this removes useless information already present in other legible files (not libs). Concerned projects are: PNG Behavior, Steven Levithan's date-time-format and Matthew Eernisse's timezone-js

I added the 2 last ones, PNG Behavior was added by David


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1693422 13f79535-47bb-0310-9956-ffa450edef68