3 patches from Christian Carlow for "FOP rendered list form widgets do not work for fields with postition > 0" https://issues.apache.org/jira/browse/OFBIZ-6354
It doesn't work because of the way the renderFormatHeaderRowOpen is called causes the FOP renderer to create more than one <fo:table-header>.

The 1st patch adds renderFormatHeaderOpen and renderFormatHeaderClose to separately generate <fo:table-header> for FOP and <thead> for html.
The 2nd patch adds the new list form widget header methods to the deprecated HtmlFormRenderer needed by EmailServices.java to handle OFBIZ-6360 until replaced with MacroFormRender.
The 3rd patch handles the other FormRenderers FOP, Text, Xml. This issue is still a work in process.

jleroux: there is this  "FIXME: this is an hack to avoid FOP rendering errors for empty lists (fo:table-body cannot be null)" in code



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1710727 13f79535-47bb-0310-9956-ffa450edef68