An updated and modified patch from Bob Morley for "ProductStore did not allow the default timezone to be set" https://issues.apache.org/jira/browse/OFBIZ-2813
The ProductStore has the defaultLocaleString and the defaultCurrencyUomId but does not have a defaultTimeZoneString.  What happens when a store is selected (from e-commerce or the POS) is that it will look to the user for a lastTimeZone and if that is not set, it would simply revert to the TimeZone.getDefault() (so the timezone based on the server).

The trouble we had was that you may have an application server that is running in a completely different timezone than what is desired for the store.

jleroux: see comments in the Jira issue for the modifications I did

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1712347 13f79535-47bb-0310-9956-ffa450edef68