3rd step for "Secure HTTP headers" https://issues.apache.org/jira/browse/OFBIZ-6766
It's recommended to not use the X-Powered-By header. 
Tomcat doc: The xpoweredBy attribute controls whether or not the X-Powered-By HTTP header is sent with each request. If sent, the value of the header contains the Servlet and JSP specification versions, the full Tomcat version (e.g. Apache Tomcat/7.0.0), the name of the JVM vendor and the version of the JVM. This header is disabled by default. This header can provide useful information to both legitimate clients and attackers. 

OFBiz users can decide to change this parameter if they want

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1719684 13f79535-47bb-0310-9956-ffa450edef68