5th step for "Secure HTTP headers" https://issues.apache.org/jira/browse/OFBIZ-6766
X-XSS-Protection

This header enables the Cross-site scripting (XSS) filter built into most recent web browsers.
It's usually enabled by default anyway, so the role of this header is to re-enable the filter for this particular website if it was disabled by the user.
This header is supported in IE 8+, and in Chrome (not sure which versions). The anti-XSS filter was added in Chrome 4. Its unknown if that version honored this header.
FireFox has still an open bug entry and "offers" only the noscript plugin.
https://wiki.mozilla.org/Security/Features/XSS_Filter
https://bugzilla.mozilla.org/show_bug.cgi?id=528661

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1720213 13f79535-47bb-0310-9956-ffa450edef68