A patch from Pierre Smits for "Enhance PartyTypeAttr with a description field" https://issues.apache.org/jira/browse/OFBIZ-6738
Currently there is no Uom selector in the terms form. And the display field in the terms list form only shows the uomId.

jleroux: I noticed that none of the *TypeAttr entities has a description field, why not adding it to them all?
Pierre: This patch addresses the issue in all applicable entities, and replaces the previously provided patch.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1743422 13f79535-47bb-0310-9956-ffa450edef68