First commit in replacing Apache ant with Gradle - OFBIZ-7534
This commit represents a pivotal change in the history of the
project in which the build system is being replaced from Apache
Ant to Gradle of which the discussion mail thread link is below.

Thanks to everyone who helped with testing and feedback

This commit delivers the following:

- add the gradle wrapper directory in gradle/* which holds
  all items necessary for gradle to run
- add the build.gradle master script
- add the settings.gradle script which declares all the subprojects
- add documentation for how to use everything in a new file called
  README.md (written in markdown)

The new script is capable of the functionality that used to exist
in Ant with some additional features including automation of eclipse
files generation and many other functions.

To complete this new feature, we need to apply the following _after_
this commit:
- change the buildbot to utilize gradle instead of ant
- delete all ant & maven scripts
- replace all local jars with remote ones from jcenter

Reference thread: http://ofbiz.markmail.org/message/vstt3wxuubmjgmqj?q=Important+Changes+to+Trunk+and+Use+of+Ant+%26+Gradle

Reference JIRA: https://issues.apache.org/jira/browse/OFBIZ-7534



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1751309 13f79535-47bb-0310-9956-ffa450edef68