A patch from Pierre Smits for "Add download definition for drivers of commonly used open source rdbms to build gradle" https://issues.apache.org/jira/browse/OFBIZ-7793
With the move to dependency mgt through gradle/gradlew the download definitions for the drivers of the most commonly used rdbms solutions were removed.
Adding these to the build.gradle file will deliver a great pleasure adoption wise through a small effort.

jleroux: This is an agreed temporary workaround while we are trying to implement interactive tasks at OFBIZ-7808

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1752843 13f79535-47bb-0310-9956-ffa450edef68