A slightly modified Taher's patch for "Load the OWASP dependency checker Gradle plugin efficiently" I reported at OFBIZ-7930
Using OWASP-DC with Gradle was impossible w/o the corresponding plugin. Unfortunately using the suggested way at https://plugins.gradle.org/plugin/org.owasp.dependencycheck loads a lot of dependencies which are only needed by OWASP-DC. So Taher wrote a specific solution to avoid downloading useless libs to a normal use of OFBiz, there it is.


jleroux: I have simply formatted the "if(" to "if ("

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1759250 13f79535-47bb-0310-9956-ffa450edef68