Implemented: New servlet filter that provides all the "controller" related logic currently available in the ContextFilter.
At the moment the filter is not used but, once the "controller" logic will be 
removed by the ContextFilter, the applications' web.xml files  will be updated 
to deploy the two filters in a chain:
* ControlFilter will filter and redirect requests
* ContextFilter will set the required objects in the contexts (session and 
request) like the delegator, dispatcher etc...
This separation of concerns will make the deplyment more flexible when 
additional filters are required.
While the logic in this new filter is functionally mostly the same of that 
currently in ContextFilter, this new implementation is more efficient and the 
behavior and configuration parameters are clearly discribed.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1760725 13f79535-47bb-0310-9956-ffa450edef68