Improved: Gradle settings for eclipse classpath
Refactor the eclipse classpath settings in build.gradle and substantially
improve it by making it denser with only one loop to remove entries. We
also added an exclusion pattern for 'dtd' directories as they were added
recently to the sourcesets block by Jacopo Capellato.

This makes the code block for eclipse settings easier to reason about


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1761519 13f79535-47bb-0310-9956-ffa450edef68