Fixes: Add WebSocket support in OFBiz (this was broken with Gradle) (OFBIZ-7073)
WebSocket support was not working after the upgrade of trunk to Gradle. 
I have slightly changed the FilterJar.java to scan only ofbiz.jar. 
It is working now. 

jleroux: Of course the trick used before with the discoverable word in jars 
filenames no longer makes sense. But since there is now only this ofbiz.jar to 
parse it's now less an issue and will not delay OFBiz start in a noticeable
 way. At least I did not notice any noticeable effect.

Thanks: Amardeep Singh Jhajj 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1761874 13f79535-47bb-0310-9956-ffa450edef68