Improves: Add WebSocket support in OFBiz (OFBIZ-7073)
With last Amardeep Singh Jhajj's patch we introduced a slow down when starting.

The simple trick is to check if we want to use websocket or not. 
By default (webSocket=false in catalina.properties) we don't and I gain 17 sec 
on my machine (13 instead of 30). Not a big deal but the log is also cleaner. 
And I guess not everybody needs websockets.

This should be documented in the wiki page to come with OFBIZ-7538. Then the 
comment for webSocket in catalina.properties should be changed accordingly

If we want to always have webSocket available, then we should start from 
Tomcat StandardJarScanFilter class and adapt it to our needs.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1763241 13f79535-47bb-0310-9956-ffa450edef68