Improved: QuoteId field visibility in Create and Update Quote process. (OFBIZ-7777)
When creating a new quote, the field quoteId is displayed even though its value is "null". Fixed the behavior adding a use-when condition to display the field when the quote already exists and to hide the field when the quote is being created.

Thanks: Montalbano Florian for the contribution and Chandan Khandelwal for reviewing and providing the updated patch.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1765012 13f79535-47bb-0310-9956-ffa450edef68