Improved: Make xssi prefix configurable (OFBIZ-9169)
In OFBIZ-5409 a prefix was added for GET ajax requests.  
It would be beneficial if the prefix could be configured instead of hard coded

Pierre Smits suggested to make the feature multi-tenant aware. I made signature 
and calls changes based on Chris Howe's patch.

Thanks: Chris Howe and Pierre Smits 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1777942 13f79535-47bb-0310-9956-ffa450edef68